requirements.txt
transformers==4.51.3
torch
gradio
Pillow
accelerate
sentencepiece
einops
torchvision
numpy
chromadb
tqdm 